Title: 합과 차의 나눗셈, Time: 205ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240816/합과 차의 나눗셈/divide-of-sum-and-sub.java
@@ -3,9 +3,9 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        double a = sc.nextDouble();
-        double b = sc.nextDouble();
+        int a = sc.nextInt();
+        int b = sc.nextInt();
 
-        System.out.printf("%.2f",((a+b)/(a-b)));
+        System.out.printf("%.2f",(double)(a+b)/(a-b));
     }
 }
